custom-elements/instantiating-custom-elements/prototype-is-interface-prototype-object.html

<!DOCTYPE html>
<html>
<head>
<title>Prototype is already an interface prototype object</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="If PROTOTYPE is already an interface prototype object for any interface object, throw a NotSupportedError and stop">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#instantiating-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();
    HTML5_ELEMENTS.forEach(function(element) {
        var obj = doc.createElement(element);        
        var name = 'x-' + element;
        //FIXME Should the type of the exception be specified here?
        assert_throws(null, function() {
            doc.registerElement(name, {prototype: obj.constructor.prototype});
        }, 'Exception should be thrown in case of attempt to register element ' +
            'if prototype is already an interface prototype object (' + 
            obj.constructor.prototype.name + ')');        
    });
}, 'Prototype is already an interface prototype object');


test(function() {
    var doc = newHTMLDocument();
    var proto = Object.create(HTMLElement.prototype);
    doc.registerElement('x-a-1', {
        prototype: proto
    });
    //FIXME Should the type of the exception be specified here?
    assert_throws(null, function() {
        doc.registerElement('x-a-1', {
            prototype: proto
        });                
    }, 'Exception shoul be thrown if custom element with already used prototype is registered');
    
}, 'Prototype is already an registered');

</script>
</body>
</html>
